refactor:  RaiseTask and TryTask builders to extend base builder classes
